# Support encoding of pointers to embedded structs