Current encoder implementation ignores embedded "pointer structs" like `*Base` in:

```go
type Base struct {
  Value string
}

type Upper struct {
  *Base
  Other bool
}
```

This PR adds support for encoding such types.

---

Paste `benchstat` results here
```
❯ ./ci.sh benchmark -d v2
Executing benchmark for v2 at /var/folders/yw/xfw4875j6xv69_2r5v52b1yc0000gn/T/tmp.lz9CXrl1y1
Preparing worktree (checking out 'v2')
HEAD is now at 34765b4 Fix unmarshaling of nested non-exported struct (#917)
/var/folders/yw/xfw4875j6xv69_2r5v52b1yc0000gn/T/tmp.lz9CXrl1y1 ~/src/go-toml
PASS
ok      github.com/pelletier/go-toml/v2 0.528s
goos: darwin
goarch: arm64
pkg: github.com/pelletier/go-toml/v2/benchmark
BenchmarkUnmarshalDataset/config-2                    97          11982439 ns/op          87.52 MB/s     5769286 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    97          12272337 ns/op          85.45 MB/s     5769337 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    99          12510009 ns/op          83.83 MB/s     5769288 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    94          12470572 ns/op          84.09 MB/s     5769284 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    98          12163871 ns/op          86.21 MB/s     5769284 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    98          12450381 ns/op          84.23 MB/s     5769335 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    99          12229840 ns/op          85.75 MB/s     5769317 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    96          12009361 ns/op          87.32 MB/s     5769323 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    99          12095854 ns/op          86.70 MB/s     5769384 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    94          12245606 ns/op          85.64 MB/s     5769363 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/canada-2                    24          45145378 ns/op          48.76 MB/s    80013612 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    26          45888026 ns/op          47.97 MB/s    80013621 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          45352332 ns/op          48.54 MB/s    80013616 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          46102220 ns/op          47.75 MB/s    80013614 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          45658970 ns/op          48.21 MB/s    80013614 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          45740437 ns/op          48.13 MB/s    80013609 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          45735265 ns/op          48.13 MB/s    80013613 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    24          45757408 ns/op          48.11 MB/s    80013610 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          45353103 ns/op          48.54 MB/s    80013614 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          45812575 ns/op          48.05 MB/s    80013610 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              91          12811796 ns/op          43.56 MB/s    34501873 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              86          12939999 ns/op          43.12 MB/s    34501890 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              82          12969068 ns/op          43.03 MB/s    34501871 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              91          12733041 ns/op          43.83 MB/s    34501812 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              88          12730549 ns/op          43.83 MB/s    34502067 B/op     177995 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              87          12749193 ns/op          43.77 MB/s    34501841 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              90          12758428 ns/op          43.74 MB/s    34501937 B/op     177995 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              88          12767396 ns/op          43.71 MB/s    34501914 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              86          12788570 ns/op          43.64 MB/s    34501790 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              87          12799251 ns/op          43.60 MB/s    34501928 B/op     177995 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  223           5365354 ns/op          82.36 MB/s    12633092 B/op      54755 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  222           5411353 ns/op          81.66 MB/s    12633504 B/op      54756 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  222           5480490 ns/op          80.63 MB/s    12633484 B/op      54756 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  211           5420278 ns/op          81.52 MB/s    12633566 B/op      54757 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  220           5367875 ns/op          82.32 MB/s    12632965 B/op      54755 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  220           5380087 ns/op          82.13 MB/s    12633223 B/op      54755 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  220           5398300 ns/op          81.86 MB/s    12632869 B/op      54754 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  220           5371905 ns/op          82.26 MB/s    12633350 B/op      54756 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  220           5403835 ns/op          81.77 MB/s    12632980 B/op      54755 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  220           5416546 ns/op          81.58 MB/s    12633022 B/op      54755 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          57112783 ns/op          47.00 MB/s    21272714 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56624383 ns/op          47.40 MB/s    21272710 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          57701471 ns/op          46.52 MB/s    21272714 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          57213358 ns/op          46.91 MB/s    21272716 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56788115 ns/op          47.26 MB/s    21272701 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          59589590 ns/op          45.04 MB/s    21272714 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      19          58964493 ns/op          45.52 MB/s    21272711 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          57316360 ns/op          46.83 MB/s    21272712 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          58259781 ns/op          46.07 MB/s    21272716 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      19          57862721 ns/op          46.39 MB/s    21272706 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            103696 ns/op          78.11 MB/s      185106 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            104333 ns/op          77.64 MB/s      185106 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            106048 ns/op          76.38 MB/s      185108 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            104168 ns/op          77.76 MB/s      185103 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            101861 ns/op          79.52 MB/s      185116 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            102526 ns/op          79.00 MB/s      185113 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            102820 ns/op          78.78 MB/s      185113 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            102209 ns/op          79.25 MB/s      185112 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            101721 ns/op          79.63 MB/s      185099 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            102418 ns/op          79.09 MB/s      185107 B/op       1305 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3667077               323.8 ns/op        33.97 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3681438               328.3 ns/op        33.50 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3639698               337.2 ns/op        32.62 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3341606               336.9 ns/op        32.66 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3591708               328.9 ns/op        33.44 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3538990               329.2 ns/op        33.41 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3683072               329.2 ns/op        33.42 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3687992               328.7 ns/op        33.47 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3685906               328.4 ns/op        33.49 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3689648               324.2 ns/op        33.93 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2603976               455.7 ns/op        24.14 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2613799               455.5 ns/op        24.15 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2514778               456.9 ns/op        24.08 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2625978               466.8 ns/op        23.57 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2469021               471.5 ns/op        23.33 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2461879               465.9 ns/op        23.61 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2549766               473.3 ns/op        23.24 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2558547               468.7 ns/op        23.47 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2586669               462.7 ns/op        23.77 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2574088               478.6 ns/op        22.98 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          40346             29712 ns/op         176.39 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          39975             30041 ns/op         174.46 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          40341             29435 ns/op         178.05 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          40458             29886 ns/op         175.37 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          40576             29444 ns/op         178.00 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          38845             30723 ns/op         170.59 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          38642             30299 ns/op         172.98 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          39476             29672 ns/op         176.63 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          40502             29811 ns/op         175.81 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          39177             29680 ns/op         176.59 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27460             43903 ns/op         119.38 MB/s       37746 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27312             43608 ns/op         120.18 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27424             43713 ns/op         119.90 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27483             43487 ns/op         120.52 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             26817             44132 ns/op         118.76 MB/s       37746 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27376             44046 ns/op         118.99 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             26563             43518 ns/op         120.43 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27436             43481 ns/op         120.54 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27466             43872 ns/op         119.46 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27460             43639 ns/op         120.10 MB/s       37746 B/op        619 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              160964              7291 ns/op          74.88 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              160708              7329 ns/op          74.50 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              160305              7270 ns/op          75.10 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              159598              7277 ns/op          75.03 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              158874              7350 ns/op          74.28 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              161134              7505 ns/op          72.75 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              161014              7264 ns/op          75.16 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              159738              7402 ns/op          73.77 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              159980              7266 ns/op          75.15 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              160560              7279 ns/op          75.01 MB/s        7441 B/op        141 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4571703               260.0 ns/op        46.16 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4568283               259.7 ns/op        46.21 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4530682               260.1 ns/op        46.13 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4569643               261.1 ns/op        45.96 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4552030               263.1 ns/op        45.60 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4260003               267.8 ns/op        44.80 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4517004               264.7 ns/op        45.33 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4471758               262.4 ns/op        45.72 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4586739               262.2 ns/op        45.77 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4525846               267.8 ns/op        44.81 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3203388               369.1 ns/op        32.51 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3289986               361.4 ns/op        33.21 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3317054               360.9 ns/op        33.25 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3248652               360.5 ns/op        33.28 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3156919               362.8 ns/op        33.08 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3331431               360.7 ns/op        33.27 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3334716               361.3 ns/op        33.22 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3205110               364.4 ns/op        32.93 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3319612               361.7 ns/op        33.18 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3234938               358.3 ns/op        33.50 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            53883             22130 ns/op          92.95 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            53301             22251 ns/op          92.45 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            54150             22521 ns/op          91.34 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            52431             22021 ns/op          93.41 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            54318             21938 ns/op          93.76 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            54172             21984 ns/op          93.57 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            54258             22439 ns/op          91.67 MB/s       28304 B/op        317 allocs/op
…